+++ pumpkin/src/plugin/api/events/player/mod.rs
@@ -10,8 +10,8 @@ pub mod player_join;
 pub mod player_leave;
 pub mod player_login;
 pub mod player_move;
-pub mod player_resource_pack_status;
-pub mod player_respawn;
+pub mod resource_pack_status;
+pub mod respawn;
 pub mod player_teleport;
 
 use std::sync::Arc;
